Mock all network calls in hw-app-eth tests
